Final commit for Email verification.
Upon registration user is sent an email with a clickable link to verify that the user controls the supplied email. Link is valid for 2 hours.
Once the link is clicked, the user can login to the system.
Once logged in the user can perform authenticated services.
If the user has forgotten the password a reset must be performed. This means a email is sent to the previously verified email-address with a clickable link that leads to a homepage where the user can change its password. if the reset is asked for several times it is the last reset link sent that is valid (for two hours).
Once reset the user can log in again.